// src/types/Enums/Specialties.ts
export enum Specialties {
  ARCHERY = 'Меткость',
  ARCHERS = 'Арбалетчики',
  GRIFFINS = 'Грифоны',
  NAVIGATION = 'Навигация',
  ESTATES = 'Поместья',
  SWORDSMEN = 'Мечники',
  BALLISTA = 'Артиллерия',
  CAVALIERS  = 'Кавалерия',
  FIRST_AID = 'Первая Помощь',
  SPEED = 'Передвижение',
  BLESS = 'Благословение',
  GOLD = 'Золото',
  WEAKNESS = 'Слабость',
  MONKS = 'Монахи',
  PRAYER = 'Молитва',
  EAGLE_EYE = 'Зоркость',
  EARTH_ELEMENTALS = 'Землянные Элементали',
  FIRE_ELEMENTALS = 'Огненные Элементали',
  WATER_ELEMENTALS = 'Водяные Элементали',
  PSYCHIC_ELEMENTALS = 'Психические Элементали',
  DISRUPTING_RAY = 'Разрушающий Луч',
  HASTE = 'Ускорение',
  MAGIC_ARROW = 'Магическая Стрела',
  BLOODLUST = 'Кровожадность',
  STONE_SKIN = 'Каменная Кожа',
  FIRE_WALL = 'Огненная Стена',
  BEHOLDERS = 'Бехолдеры',
  MINOTAURS = 'Минотавры',
  LOGISTICS = 'Логистика',
  HARPIES = 'Гарпии',
  TROGLODYTES = 'Троглодиты',
  MANTICORES = 'Мантикоры',
  RESURRECTION = 'Воскрешение',
  METEOR_SHOWER = 'Метеоритный Дождь',
  MYSTICISM = 'Мистицизм',
  SORCERY = 'Колдовство',
  CRYSTAL = 'Кристаллы',
  GORGONS = 'Горгоны',
  WYVERNS = 'Виверны',
  BASILISKS = 'Василиски',
  GNOLLS = 'Гноллы',
  SERPENT_FLIES = 'Летучие Змеи',
  DEFENSE = 'Защита',
  LIZARDMEN = 'Ящеры',
  INTELLIGENCE = 'Интеллект',
  GOGS = 'Гоги',
  HELL_HOUNDS = 'Церберы',
  IMPS = 'Бесы',
  DEMONS = 'Демоны',
  PIT_FIENDS = 'Порождения Зла',
  EFREETS = 'Ифриты',
  SULFUR = 'Сера',
  FIREBALL = 'Огненный Шар',
  INFERNO = 'Инферно',
  WIGHTS = 'Призраки',
  SKELETONS = 'Скелеты',
  NECROMANCY = 'Некромантия',
  Liches = 'Личи',
  WALKING_DEAD = 'Ходячие Мертвецы',
  BLACK_KNIGHTS = 'Рыцари Смерти',
  VAMPIRES = 'Вампиры',
  DEATH_RIPPLE = 'Волна Смерти',
  ANIMATE_DEAD = 'Оживление Мертвецов',
  UNICORNS = 'Единороги',
  ELFS = 'Ельфы',
  DENDROIDS = 'Дендроиды',
  RESISTANCE = 'Сопротивление',
  DWARVES = 'Гномы',
  PEGASI = 'Пегасы',
  ICE_BOLT = 'Ледяная Стрела',
  SLAYER = 'Убийца',
  FORTUNE = 'Удача',
  CURE = 'Лечение',
  OFFENSE = 'Атака',
  GOBLINS = 'Гоблины',
  ORCS = 'Орки',
  OGRES = 'Огры',
  ROCS = 'Птицы Рух',
  WOLF_RIDERS = 'Наездники На Волках',
  CYCLOPES = 'Циклопы',
  GEMS = 'Самоцветы',
  PRECISION = 'Точность',
  NAGAS = 'Наги',
  GENIES = 'Джинны',
  GOLEMS = 'Големы',
  GARGOYLES = 'Гаргульи',
  MERCURY = 'Ртуть',
  HYPNOTIZE = 'Гипноз',
  CHAIN_LIGHTNING = 'Цепная Молния',
  MAGI = 'Маги',
}

export enum SpecialtyPics {
  ARCHERY = ''
}